backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
python-docx==0.8.11
python-pptx==0.6.23
google-generativeai==0.3.0
openai==1.3.0
requests==2.31.0
redis==5.0.0
aioredis==2.0.1
python-multipart==0.0.6
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
cors==1.0.1
pillow==10.0.1
jinja2==3.1.2
cryptography==41.0.7
